Codex/requirements.txt
langchain
langchain-openai
langgraph
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
aiohttp
gitpython
slack-sdk
jira
PyGithub
flake8
bandit
radon
tenacity
python-json-logger
orjson
sqlalchemy
alembic
cryptography
openai
pytest
requests-mock
langchain-core
azure-identity
